+++ korr/__init__.py
@@ -6,3 +6,5 @@ from .flatten import flatten
 from .corrgram import corrgram
 from .corr_vs_pval import corr_vs_pval
 from .bracket_pval import bracket_pval
+from .confusion import confusion
+from .mcc import mcc, confusion_to_mcc